requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.10.8
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cx-Oracle==8.2.1
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
entrypoints==0.4
et-xmlfile==1.0.1
executing==0.8.3
filterpy==1.4.5
greenlet==1.1.1
hubTools==1.0.16
ipykernel==6.9.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
nest-asyncio==1.5.4
numexpr==2.7.3
numpy==1.21.1
olefile==0.46
openpyxl==3.0.9
pandas==1.3.1
parse==1.19.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.3
prompt-toolkit==3.0.19
psutil==5.9.0
pure-eval==0.2.2
PyCygNet==0.0.6
Pygments==2.9.0
pyodbc==4.0.31
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32==302
pyzmq==22.1.0
schedule==1.1.0
scipy==1.7.1
setuptools==49.6.0.post20210108
six==1.16.0
SQLAlchemy==1.4.22
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
wheel==0.37.0
wincertstore==0.2
# -e git+https://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@github.com/tygast/hubTools.git#egg=hubTools
# -e git+https://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@git.eogresources.com/tgaston/PyCygNet.git#egg=PyCygNet